refactor: move neqsim mapping to neqsim wrapper
