+++ src/ecalc/libraries/neqsim/neqsim_ecalc_wrapper/__init__.py
@@ -3,7 +3,7 @@ from neqsim_ecalc_wrapper.java_service import start_server
 java_gateway = start_server()
 neqsim = java_gateway.jvm.neqsim
 
-from neqsim_ecalc_wrapper.thermo import NeqsimEoSModelType, NeqsimFluid
+from neqsim_ecalc_wrapper.thermo import NeqsimFluid
 
 
 def methods(check_class):
