+++ src/ecalc/libraries/neqsim/neqsim_ecalc_wrapper/thermo.py
@@ -2,19 +2,26 @@ from __future__ import annotations
 
 import os
 from dataclasses import dataclass
-from enum import Enum
 from functools import lru_cache
 from pathlib import Path
 from typing import Dict, List, Tuple, Union
 
+from libecalc import dto
 from libecalc.common.capturer import Capturer
 from libecalc.common.logger import logger
+from libecalc.dto.types import EoSModel
 from py4j.protocol import Py4JJavaError
 from pydantic import BaseModel
 
 from neqsim_ecalc_wrapper import neqsim
 from neqsim_ecalc_wrapper.components import COMPONENTS
 from neqsim_ecalc_wrapper.exceptions import NeqsimPhaseError
+from neqsim_ecalc_wrapper.mappings import (
+    NeqsimEoSModelType,
+    _map_fluid_component_from_neqsim,
+    map_eos_model_to_neqsim,
+    map_fluid_composition_to_neqsim,
+)
 
 STANDARD_TEMPERATURE_KELVIN = 288.15
 STANDARD_PRESSURE_BARA = 1.01325
@@ -23,13 +30,6 @@ NEQSIM_MIXING_RULE = 2
 
 ThermodynamicSystem = neqsim.thermo.system.SystemEos
 ThermodynamicOperations = neqsim.thermodynamicOperations.ThermodynamicOperations
-
-
-class NeqsimEoSModelType(Enum):
-    SRK = neqsim.thermo.system.SystemSrkEos
-    PR = neqsim.thermo.system.SystemPrEos
-    GERG_SRK = neqsim.thermo.system.SystemSrkEos
-    GERG_PR = neqsim.thermo.system.SystemPrEos
 
 
 class NeqsimFluidComponent(BaseModel):
@@ -120,10 +120,10 @@ class NeqsimFluid:
     @classmethod
     def create_thermo_system(
         cls,
-        composition: Dict[str, float],
+        composition: dto.FluidComposition,
         temperature_kelvin: float = STANDARD_TEMPERATURE_KELVIN,
         pressure_bara: float = STANDARD_PRESSURE_BARA,
-        eos_model: NeqsimEoSModelType = NeqsimEoSModelType.SRK,
+        eos_model: EoSModel = EoSModel.SRK,
         mixing_rule: int = NEQSIM_MIXING_RULE,
     ) -> NeqsimFluid:
         """Initiates a NeqsimFluid that wraps both a Neqsim thermodynamic system and operations.
@@ -142,6 +142,9 @@ class NeqsimFluid:
         :return:
         """
         use_gerg = "gerg" in eos_model.name.lower()
+
+        composition = map_fluid_composition_to_neqsim(fluid_composition=composition)
+        eos_model = map_eos_model_to_neqsim(eos_model)
         non_existing_components = [
             component for component, value in composition.items() if (component not in COMPONENTS) and (value > 0.0)
         ]
@@ -280,24 +283,9 @@ class NeqsimFluid:
         mass_rate_stream_2: float,
         pressure: float,
         temperature: float,
-        eos_model: NeqsimEoSModelType = NeqsimEoSModelType.SRK,
-    ) -> Tuple[Dict, NeqsimFluid]:
+        eos_model: EoSModel = EoSModel.SRK,
+    ) -> Tuple[dto.FluidComposition, NeqsimFluid]:
         """Mixing two streams (NeqsimFluids) with same pressure and temperature."""
-        if stream_1 is None:
-            return (
-                {},  # Fixme: Need to return composition dict
-                stream_2.set_new_pressure_and_temperature(
-                    new_pressure_bara=pressure, new_temperature_kelvin=temperature
-                ),
-            )
-        if stream_2 is None:
-            return (
-                {},  # Fixme: Need to return composition dict
-                stream_1.set_new_pressure_and_temperature(
-                    new_pressure_bara=pressure,
-                    new_temperature_kelvin=temperature,
-                ),
-            )
 
         composition_dict: Dict[str, float] = {}
 
@@ -316,10 +304,14 @@ class NeqsimFluid:
                 else:
                     composition_dict[composition_name] = composition_moles
 
+        ecalc_fluid_composition = dto.FluidComposition.parse_obj(
+            {_map_fluid_component_from_neqsim[key]: value for (key, value) in composition_dict.items()}
+        )
+
         return (
-            composition_dict,
+            ecalc_fluid_composition,
             NeqsimFluid.create_thermo_system(
-                composition=composition_dict,
+                composition=ecalc_fluid_composition,
                 temperature_kelvin=temperature,
                 pressure_bara=pressure,
                 eos_model=eos_model,
